requirements.txt
requests
statsmodels
numpy
PyYAML
pandas
matplotlib
bokeh
scikit-learn
openpyxl
pyproj